[#259] Provide device registry in hono-service-base.
Introduced new parent class and moved logic.
