+++ services/device-registry/src/main/java/org/eclipse/hono/deviceregistry/SimpleDeviceRegistryServer.java
@@ -12,127 +12,16 @@
 
 package org.eclipse.hono.deviceregistry;
 
-import io.vertx.proton.*;
-import org.eclipse.hono.connection.ConnectionFactory;
-import org.eclipse.hono.service.amqp.AmqpServiceBase;
-import org.eclipse.hono.service.amqp.Endpoint;
-import org.eclipse.hono.service.auth.AuthenticationConstants;
-import org.eclipse.hono.util.Constants;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
+import org.eclipse.hono.service.registration.BaseDeviceRegistryServer;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import io.vertx.core.AsyncResult;
-import io.vertx.core.json.JsonObject;
-import io.vertx.ext.healthchecks.HealthCheckHandler;
-import io.vertx.ext.healthchecks.Status;
-
-import java.util.Objects;
-import java.util.UUID;
-
 
 /**
- * A server that implements Hono's <a href="https://www.eclipse.org/hono/api/Device-Registration-API/">Device Registration API</a> and
- * <a href="https://www.eclipse.org/hono/api/Credentials-API/">Credentials API</a> as the default implementation of Hono's device registry.
- *
- * TODO: add metrics.
+ * Spring component that serves as the default implementation of Hono's device registry.
  */
 @Component
 @Scope("prototype")
-public final class SimpleDeviceRegistryServer extends AmqpServiceBase<DeviceRegistryConfigProperties> {
+public final class SimpleDeviceRegistryServer extends BaseDeviceRegistryServer<DeviceRegistryConfigProperties> {
 
-    private ConnectionFactory authenticationService;
-
-    /**
-     * Sets the factory to use for creating an AMQP 1.0 connection to
-     * the Authentication service.
-     * 
-     * @param factory The factory.
-     * @throws NullPointerException if factory is {@code null}.
-     */
-    @Autowired
-    @Qualifier(AuthenticationConstants.QUALIFIER_AUTHENTICATION)
-    public void setAuthenticationServiceConnectionFactory(final ConnectionFactory factory) {
-        authenticationService = Objects.requireNonNull(factory);
-    }
-
-    @Override
-    protected void onRemoteConnectionOpen(final ProtonConnection connection) {
-        connection.setContainer(String.format("Hono-DeviceRegistry-%s:%d", getBindAddress(), getPort()));
-        setRemoteConnectionOpenHandler(connection);
-    }
-
-    @Override
-    protected void onRemoteConnectionOpenInsecurePort(final ProtonConnection connection) {
-        connection.setContainer(String.format("Hono-DeviceRegistry-%s:%d", getInsecurePortBindAddress(), getInsecurePort()));
-        setRemoteConnectionOpenHandler(connection);
-    };
-
-    private void setRemoteConnectionOpenHandler(final ProtonConnection connection) {
-        connection.sessionOpenHandler(remoteOpenSession -> handleSessionOpen(connection, remoteOpenSession));
-        connection.receiverOpenHandler(remoteOpenReceiver -> handleReceiverOpen(connection, remoteOpenReceiver));
-        connection.senderOpenHandler(remoteOpenSender -> handleSenderOpen(connection, remoteOpenSender));
-        connection.disconnectHandler(this::handleRemoteDisconnect);
-        connection.closeHandler(remoteClose -> handleRemoteConnectionClose(connection, remoteClose));
-        connection.openHandler(remoteOpen -> {
-            LOG.info("client [container: {}, user: {}] connected", connection.getRemoteContainer(), Constants.getClientPrincipal(connection).getName());
-            connection.open();
-            // attach an ID so that we can later inform downstream components when connection is closed
-            connection.attachments().set(Constants.KEY_CONNECTION_ID, String.class, UUID.randomUUID().toString());
-        });
-    }
-
-
-    private void handleSessionOpen(final ProtonConnection con, final ProtonSession session) {
-        LOG.info("opening new session with client [{}]", con.getRemoteContainer());
-        session.closeHandler(sessionResult -> {
-            if (sessionResult.succeeded()) {
-                sessionResult.result().close();
-            }
-        }).open();
-    }
-
-    /**
-     * Invoked when a client closes the connection with this server.
-     *
-     * @param con The connection to close.
-     * @param res The client's close frame.
-     */
-    private void handleRemoteConnectionClose(final ProtonConnection con, final AsyncResult<ProtonConnection> res) {
-        if (res.succeeded()) {
-            LOG.info("client [{}] closed connection", con.getRemoteContainer());
-        } else {
-            LOG.info("client [{}] closed connection with error", con.getRemoteContainer(), res.cause());
-        }
-        con.close();
-        con.disconnect();
-    }
-
-    private void handleRemoteDisconnect(final ProtonConnection connection) {
-        LOG.info("client [{}] disconnected", connection.getRemoteContainer());
-        connection.disconnect();
-    }
-
-    @Override
-    public void registerReadinessChecks(final HealthCheckHandler handler) {
-        for (Endpoint ep : endpoints()) {
-            ep.registerReadinessChecks(handler);
-        }
-        handler.register("authentication-service-connection", status -> {
-            if (authenticationService == null) {
-                status.complete(Status.KO(new JsonObject().put("error", "no connection factory set for Authentication service")));
-            } else {
-                LOG.debug("checking connection to Authentication service");
-                authenticationService.connect(null, null, null, s -> {
-                    if (s.succeeded()) {
-                        s.result().close();
-                        status.complete(Status.OK());
-                    } else {
-                        status.complete(Status.KO(new JsonObject().put("error", "cannot connect to Authentication service")));
-                    }
-                });
-            }
-        });
-    }
 }
